registration forms for both simple user and supplier
